Fix facilitator query bug
There was a bug in the facilitator getAll query where we were querying
on a salesforce label instead of a name. Salesforce stupidly calls
object labels `Name` in a SOQL query, and calls object names
`DeveloperName`. So when case requested the label to be changed for the
salesforce frontend, it broke the query.